requirements.txt
anyio==3.6.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
diffusers==0.2.2
fastapi==0.79.0
filelock==3.8.0
future==0.18.2
h11==0.13.0
huggingface==0.0.1
huggingface-hub==0.8.1
idna==3.3
importlib-metadata==4.12.0
numpy==1.23.2
packaging==21.3
Pillow==9.2.0
pydantic==1.9.2
pyngrok==5.1.0
pyparsing==3.0.9
python-dotenv==0.20.0
PyYAML==6.0
regex==2022.7.25
requests==2.28.1
sniffio==1.2.0
starlette==0.19.1
tokenizers==0.12.1
torch==1.12.1
tqdm==4.64.0
transformers==4.21.1
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
zipp==3.8.1
